Moved cart and game purchase implementation to its own folder.
